filter and some delete

+++ CREATE_TABLES.sql
@@ -3,14 +3,14 @@ CREATE SCHEMA WILDLIFE_SCHEMA;
 USE WILDLIFE_SCHEMA;
 
 CREATE TABLE Location (
-  Latitude FLOAT NOT NULL,
-  Longitude FLOAT NOT NULL,
+  Latitude DECIMAL(6,4) NOT NULL,
+  Longitude DECIMAL(7,4) NOT NULL,
   LocationName VARCHAR(255),
   LocationType VARCHAR(50),
   Country VARCHAR(50),
-  Area FLOAT,
+  Area DECIMAL(10,2),
   Climate VARCHAR(50),
-  Elevation FLOAT,
+  Elevation DECIMAL(6,2),
   PRIMARY KEY (Latitude, Longitude)
 );
 
@@ -19,8 +19,8 @@ CREATE TABLE Habitat (
   HabitatType VARCHAR(50),
   ConservationStatus VARCHAR(50),
   DegradationLevel VARCHAR(50),
-  Latitude FLOAT,
-  Longitude FLOAT,
+  Latitude DECIMAL(6,4),
+  Longitude DECIMAL(7,4),
   FOREIGN KEY (Latitude, Longitude) REFERENCES Location (Latitude, Longitude) ON DELETE CASCADE
 );
